requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Supabase
supabase==2.9.1

# JWT & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Templates
jinja2==3.1.4

# HTML Sanitization (XSS protection)
bleach==6.1.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
